Move showing lock screen to BaseActivity #72
