+++ app/src/main/java/bitcoin/wallet/BaseActivity.kt
@@ -1,0 +1,29 @@
+package bitcoin.wallet
+
+import android.content.Intent
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import android.view.WindowManager
+import bitcoin.wallet.core.App
+import bitcoin.wallet.modules.main.UnlockActivity
+
+abstract class BaseActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        window.setFlags(WindowManager.LayoutParams.FLAG_SECURE, WindowManager.LayoutParams.FLAG_SECURE)
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        if (App.promptPin) {
+            startActivity(Intent(this, UnlockActivity::class.java))
+            return
+        }
+
+        App.promptPin = false
+    }
+
+}
