+++ app/src/main/java/bitcoin/wallet/LauncherActivity.kt
@@ -5,14 +5,13 @@ import android.content.Intent
 import android.os.Bundle
 import android.security.keystore.KeyPermanentlyInvalidatedException
 import android.security.keystore.UserNotAuthenticatedException
-import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import bitcoin.wallet.core.managers.Factory
 import bitcoin.wallet.core.security.EncryptionManager
 import bitcoin.wallet.modules.main.MainModule
 import java.security.UnrecoverableKeyException
 
-class LauncherActivity : AppCompatActivity() {
+class LauncherActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
